login, logout, cookie, js session, upload doc, upload image, update laporan

+++ app/Controllers/Api/V1/Kunjungan.php
@@ -4,6 +4,7 @@ use App\Controllers\BaseController;
 use App\Models\KunjunganModel;
 use App\Models\UpiModel;
 use App\Models\PembinaMutuJabatanModel;
+use CodeIgniter\I18n\Time;
 
 class Kunjungan extends BaseController
 {
@@ -36,12 +37,27 @@ class Kunjungan extends BaseController
 			]
 		)
 	);
+	protected $month = array(
+		'1' 	=> 	'Jan',
+		'2' 	=> 	'Feb',
+		'3' 	=> 	'Mar',
+		'4' 	=> 	'Apr',
+		'5' 	=> 	'Mei',
+		'6' 	=> 	'Jun',
+		'7' 	=> 	'Jul',
+		'8' 	=> 	'Agu',
+		'9' 	=> 	'Sep',
+		'10' 	=> 	'Okt',
+		'11' 	=> 	'Nov',
+		'12' 	=> 	'Des'
+	);
+	protected $userSession;
 
 	function __construct()
 	{
 		$this->KunjunganModel = new KunjunganModel();
 		$this->validation = \Config\Services::validation();
-		$this->db = db_connect();
+		$this->userSession = session('auth');
 	}
 
 	public function GetAll()
@@ -54,8 +70,8 @@ class Kunjungan extends BaseController
 
 		$q = $req->getGet();
 
-		$page = (isset($q['page'])) ? $q['page'] : 1;
-		$limit = (isset($q['limit'])) ? $q['limit'] : 1;
+		$page = isset($q['page']) && filter_var($q['page'], FILTER_VALIDATE_INT) ? (int)$q['page'] : 1;
+		$limit = isset($q['limit']) && filter_var($q['limit'], FILTER_VALIDATE_INT) ? (int)$q['limit'] : 1;
 		$offset = ($page - 1) * $limit;
 
 		// filter injection here
@@ -94,14 +110,19 @@ class Kunjungan extends BaseController
 			->join($joinTables[0], $joinCondition[0])
 			->join($joinTables[1], $joinCondition[1])
 			->join($joinTables[2], $joinCondition[2])
-			->where($whereQuery)->orderBy('tbl_kunjungan.id', 'desc')->findAll($limit, $offset);
+			->where($whereQuery)
+			->orderBy('tbl_kunjungan.id', 'desc')
+			->findAll($limit, $offset);
 
 		$countQuery = $this->KunjunganModel
 			->join($joinTables[0], $joinCondition[0])
 			->join($joinTables[1], $joinCondition[1])
 			->join($joinTables[2], $joinCondition[2])
-			->where($whereQuery)->selectCount('tbl_kunjungan.id')->find();
+			->where($whereQuery)
+			->selectCount('tbl_kunjungan.id')
+			->find();
 		$count = (int)$countQuery[0]['id'];
+
 
 		$pageAvailable = ceil((int)$count/(int)$limit);
 
@@ -128,6 +149,11 @@ class Kunjungan extends BaseController
 					} else {
 						$resp[$key]['unit_kerja_terakhir'] = 'Tidak ada jabatan aktif';
 					}
+
+					$tglKunjungan = Time::parse($value['tanggal_kunjungan']);
+
+					$resp[$key]['tanggal_kunjungan'] = $tglKunjungan->getDay().'-'.$this->month[$tglKunjungan->getMonth()].'-'.$tglKunjungan->getYear();
+					$resp[$key]['catatan'] = html_entity_decode($value['catatan']);
 				}
 			}
 
@@ -183,7 +209,29 @@ class Kunjungan extends BaseController
 			return ResponseNotAllowed();
 		}
 
-		$resp = $this->KunjunganModel->find($id);
+		$selectQuery = 'tbl_kunjungan.*';
+		$joinTables = array();
+		$joinCondition = array();
+
+		$selectQuery .= ', tbl_pembina_mutu.nama_lengkap as nama_pembina_mutu';
+		array_push($joinTables, 'tbl_pembina_mutu');
+		array_push($joinCondition, 'tbl_pembina_mutu.id = tbl_kunjungan.pembina_mutu_id');
+
+		$selectQuery .= ', tbl_upi.nama_perusahaan as nama_upi';
+		array_push($joinTables, 'tbl_upi');
+		array_push($joinCondition, 'tbl_upi.id = tbl_kunjungan.upi_id');
+
+		$selectQuery .= ', tbl_location.name as nama_provinsi';
+		array_push($joinTables, 'tbl_location');
+		array_push($joinCondition, 'tbl_upi.provinsi = tbl_location.id');
+
+		$resp = $this->KunjunganModel->select($selectQuery)
+			->join($joinTables[0], $joinCondition[0])
+			->join($joinTables[1], $joinCondition[1])
+			->join($joinTables[2], $joinCondition[2])
+			->orderBy('tbl_kunjungan.id', 'desc')->find($id);
+
+		$resp['catatan'] = html_entity_decode($resp['catatan']);
 
 		$transformed = array(
 			'data' => $resp
@@ -218,14 +266,15 @@ class Kunjungan extends BaseController
 			return ResponseNotFound();
 		}
 
-		$pembinaMutuId = 10;
+		$pembinaMutuId = (isset($this->userSession['pmid'])) ? (int)$this->userSession['pmid'] : 0;
 
 		$insert = array_merge(
 			$reqArray,
 			array(
-				'pembina_mutu_id'	=> $pembinaMutuId
+				'pembina_mutu_id' => $pembinaMutuId
 			)
 		);
+
 
 		$this->KunjunganModel->save($insert);
 
@@ -275,13 +324,78 @@ class Kunjungan extends BaseController
 		return ResponseOK(array( 'message' => 'kunjungan updated' ));
 	}
 
-	public function delete($id = 0) {
+	public function delete($id = 0)
+	{
 		if ($id < 1) {
 			return ResponseNotFound();
+		}
+
+		$req = $this->request;
+
+		if ($req->getMethod(TRUE) !== 'DELETE') {
+			return ResponseNotAllowed();
 		}
 
 		$deleted = $this->KunjunganModel->delete($id);
 
 		return ResponseOK(array( 'message' => 'kunjungan deleted' ));
+	}
+
+	public function GetFileByKunjunganId($id = 0)
+	{
+		if ($id < 1) {
+			return ResponseNotFound();
+		}
+
+		$req = $this->request;
+
+		if ($req->getMethod(TRUE) !== 'GET') {
+			return ResponseNotAllowed();
+		}
+
+		$q = 'SELECT fk.id, fk.file_upload_id, fu.upload_path, fu.file_name, fu.file_type
+            FROM tbl_file_kunjungan fk
+            JOIN tbl_file_upload fu
+            ON fk.file_upload_id = fu.id
+			AND fk.kunjungan_id = ?
+			AND fk.file_type = ?
+			AND (fk.deleted_at IS NULL OR fk.deleted_at = ?)
+			AND (fu.deleted_at IS NULL OR fu.deleted_at = ?)
+        ';
+
+		$resp['images'] = $this->db->query($q, [$id, 'image', '0000-00-00 00:00:00', '0000-00-00 00:00:00'])->getResultArray();
+		$resp['documents'] = $this->db->query($q, [$id, 'document', '0000-00-00 00:00:00', '0000-00-00 00:00:00'])->getResultArray();
+
+        return ResponseOK(array( 'data' => $resp ));
+	}
+
+	public function deleteFile($id = 0)
+	{
+		if ($id < 1) {
+			return ResponseNotFound();
+		}
+
+		$req = $this->request;
+
+		if ($req->getMethod(TRUE) !== 'DELETE') {
+			return ResponseNotAllowed();
+		}
+
+		$tfk = $this->db->table('tbl_file_kunjungan');
+		$tfu = $this->db->table('tbl_file_upload');
+
+		$getfile = $tfk->select('tbl_file_upload.id, tbl_file_upload.upload_path')
+			->join('tbl_file_upload', 'tbl_file_upload.id = tbl_file_kunjungan.file_upload_id')
+			->where('tbl_file_kunjungan.id', $id)
+			->get()->getRow();
+
+		if ((int)$getfile->id) {
+			$tfk->where('id', $id)->update(array('deleted_at' => date('Y-m-d H:i:s')));
+			$tfu->where('id', (int)$getfile->id)->update(array('deleted_at' => date('Y-m-d H:i:s')));
+			$delpath = './'.$getfile->upload_path;
+			@unlink($delpath);
+		}
+
+		return ResponseOK(array( 'message' => 'File berhasil dihapus' ));
 	}
 }
